Added grade functionality for front/back end
Made a selector where you can select between valid grades. When
a grade is selected it calls on the update grade DB function
that sends a post, updating the user database. For
session grade save changes, it uses the same function
the way courses are saved with a small change for grade saves.
Updated the course structure to include grade.
Small update to popup to show course description now.